feat: statistics와 관련된 method들을 dogs module에서 분리해서 순환 종속성 해결
dogs service에서 statistics를 구현하려고 하니 순환 종속성 문제가 발생했고, NestJS 공식 문서에 안내되어 있는 것처럼 forwardRef를 사용해보았으나 해결되지 않았다.
설계 자체에 문제가 있다고 보고 아예 statistics와 관련된 부분을 따로 분리해 dogs와 journals간의 순환 종속성을 제거했다.
