+++ backend/src/dogs/dogs.controller.ts
@@ -1,22 +1,9 @@
-import {
-    Body,
-    Controller,
-    Delete,
-    Get,
-    HttpCode,
-    Param,
-    ParseIntPipe,
-    Patch,
-    Post,
-    Query,
-    UseGuards,
-} from '@nestjs/common';
+import { Body, Controller, Delete, Get, HttpCode, Param, ParseIntPipe, Patch, Post, UseGuards } from '@nestjs/common';
 import { AccessTokenPayload } from '../auth/token/token.service';
 import { User } from '../users/decorators/user.decorator';
 import { DogsService } from './dogs.service';
 import { DogDto } from './dto/dog.dto';
 import { AuthDogGuard } from './guards/authDog.guard';
-import { DateValidationPipe } from './pipes/dateValidation.pipe';
 
 export type DogProfile = {
     id: number;
@@ -55,21 +42,6 @@ export class DogsController {
     @UseGuards(AuthDogGuard)
     async getOneProfile(@Param('id', ParseIntPipe) dogId: number) {
         return this.dogsService.getProfile(dogId);
-    }
-
-    @Get('/:id(\\d+)/statistics')
-    @UseGuards(AuthDogGuard)
-    async getDogStatistics(
-        @User() { userId }: AccessTokenPayload,
-        @Param('id', ParseIntPipe) dogId: number,
-        @Query('date', DateValidationPipe) date: string
-    ) {
-        return await this.dogsService.getDogStatistics(userId, dogId, date);
-    }
-
-    @Get('/statistics')
-    async getDogsStatistics(@User() { userId }: AccessTokenPayload) {
-        return await this.dogsService.getDogsStatistics(userId);
     }
 
     @Get('/walk-available')
