+++ backend/src/dogs/dogs.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { makeSubObjectsArray } from 'src/utils/manipulate.util';
 import { FindOptionsWhere, In, UpdateResult } from 'typeorm';
 import { QueryDeepPartialEntity } from 'typeorm/query-builder/QueryPartialEntity';
@@ -13,7 +13,6 @@ import { UsersService } from '../users/users.service';
 import { DogProfile } from './dogs.controller';
 import { Dogs } from './dogs.entity';
 import { DogsRepository } from './dogs.repository';
-import { DogStatisticDto } from './dto/dog-statistic.dto';
 import { DogDto } from './dto/dog.dto';
 
 @Injectable()
@@ -126,62 +125,5 @@ export class DogsService {
         const availableDogList = await this.dogsRepository.find({ id: In(ownDogIds), isWalking: false });
         const availableDogProfileList = this.makeProfileList(availableDogList);
         return availableDogProfileList;
-    }
-
-    private makeStatisticData(
-        dogProfiles: DogProfile[],
-        recommendedWalkAmount: number[],
-        todayWalkAmount: number[],
-        weeklyWalks: number[][]
-    ): DogStatisticDto[] {
-        const result: DogStatisticDto[] = [];
-        for (let i = 0; i < dogProfiles.length; i++) {
-            result.push({
-                id: dogProfiles[i].id,
-                name: dogProfiles[i].name,
-                profilePhotoUrl: dogProfiles[i].profilePhotoUrl,
-                recommendedWalkAmount: recommendedWalkAmount[i],
-                todayWalkAmount: todayWalkAmount[i],
-                weeklyWalks: weeklyWalks[i],
-            });
-        }
-        return result;
-    }
-
-    async getDogStatistics(userId: number, dogId: number, date: string) {
-        const dog = await this.findOne({ id: dogId });
-        // this.journalsService.getDogMonthlyJournals(userId, dogId, date);
-
-        return { recommendedWalkAmount: dog.breed.recommendedWalkAmount };
-    }
-
-    async getDogsStatistics(userId: number): Promise<DogStatisticDto[]> {
-        const ownDogIds = await this.usersService.getOwnDogsList(userId);
-        const dogWalkDayIds = await this.getRelatedTableIdList(ownDogIds, 'walkDayId');
-        const dailyWalkTimeIds = await this.getRelatedTableIdList(ownDogIds, 'dailyWalkTimeId');
-        const breedIds = await this.getRelatedTableIdList(ownDogIds, 'breedId');
-
-        const dogProfiles = await this.getProfileList({ id: In(ownDogIds) });
-        const recommendedWalkAmount = await this.breedService.getRecommendedWalkAmountList(breedIds);
-        const todayWalkAmount = await this.dailyWalkTimeService.getWalkTimeList(dailyWalkTimeIds);
-        const weeklyWalks = await this.dogWalkDayService.getWalkDayList(dogWalkDayIds);
-
-        const length = ownDogIds.length;
-        const mismatchedLengths = [
-            dogProfiles.length !== length && `dogProfiles.length = ${dogProfiles.length}`,
-            recommendedWalkAmount.length !== length && `recommendedWalkAmount.length = ${recommendedWalkAmount.length}`,
-            todayWalkAmount.length !== length && `todayWalkAmount.length = ${todayWalkAmount.length}`,
-            weeklyWalks.length !== length && `weeklyWalks.length = ${weeklyWalks.length}`,
-        ].filter(Boolean);
-        if (mismatchedLengths.length > 0) {
-            const error = new NotFoundException(`Data missing or mismatched for dogId: ${ownDogIds}.`);
-            this.logger.error(
-                `Data missing or mismatched for dogId: ${ownDogIds}. Expected length: ${length}. Mismatched data: ${mismatchedLengths.join(', ')}`,
-                error.stack ?? 'No stack'
-            );
-            throw error;
-        }
-
-        return this.makeStatisticData(dogProfiles, recommendedWalkAmount, todayWalkAmount, weeklyWalks);
     }
 }
